refactor: move deployenv to module
