+++ cmd/agent/internal/deployenv/dto.go
@@ -1,0 +1,10 @@
+package deployenv
+
+type ModifyDeploymentEnvRequest struct {
+	ProjectId string `swaggerignore:"true"`
+	Name      string `json:"name" validate:"required,kebabnum"`
+}
+
+type ListDeploymentEnvResponse struct {
+	Data []string `json:"data"`
+}
